Moved most debugging messages to console instead of log. Added StockManager class Items now update StockManager quantities automatically based on creation, destruction and faction ownership changes Fixed construction destruction so that items used in the building get released properly Peaceful wild animals now flee when a creature of differing faction gets near Removed commandline argument support Added config.ini file support Slightly optimized GameMap::Draw()

+++ Goblin Camp/src/Map.cpp
@@ -20,7 +20,7 @@ GameMap* GameMap::Inst() {
 }
 
 float GameMap::getWalkCost(int x0, int y0, int x1, int y1, void *data) const {
-	if (Walkable(x1,y1)) return tileMap[x0][y0].moveCost();
+	if (Walkable(x1,y1)) return (float)tileMap[x0][y0].moveCost();
 	return 0.0f;
 }
